Modulo-9-EBAC
Module 9 EBAC - Creating a Game: Pong - Basic Moves

A scene containing two characters (enemy and player), represented by the right and left paddles, and a moving ball capable of colliding with the paddles and changing its direction.
The game has a UI for scoring points. Points are scored when the ball passes the paddles and touches the end of the screen.
The game restarts after a defeat is declared if the enemy scores 3 points, or a victory if the player scores 3 points.